pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "protembedder"
version = "0.1.0"
description = "Extract protein embeddings from FASTA files using ESM-2 models"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "ProtEmbedder Contributors"},
]
keywords = ["protein", "embeddings", "ESM", "bioinformatics", "deep-learning", "pytorch"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=1.12.0",
    "fair-esm>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
]

[project.scripts]
protembedder = "protembedder.cli:main"

[project.urls]
Homepage = "https://github.com/youruser/protembedder"
Issues = "https://github.com/youruser/protembedder/issues"

[tool.setuptools.packages.find]
include = ["protembedder*"]
